Glitch Garden Remastered 12/2/2022

+++ Glitch Garden Remastered/Assets/Scripts/Level.cs
@@ -80,7 +80,7 @@ public class Level : SerializedScriptableObject {
     }
     
     public float GetDeviationPercent() {
-        return deviationPercent / 100;
+        return (float)deviationPercent / 100;
     }
     
     public Season GetSeason() {
